:art: 쿼리 호출부 훅으로 분리
