+++ src/hooks/useFetchAllPoses.tsx
@@ -1,0 +1,9 @@
+import { useQuery } from '@tanstack/react-query';
+import { poseService } from '../api/poseService';
+
+export default function useFetchAllPoses() {
+  return useQuery({
+    queryKey: ['allPoses'],
+    queryFn: () => poseService.fetchAllPoses(),
+  });
+}
